+++ storage/framework/views/eda0072958ea7171a69812a75031ad8084412ff9.php
@@ -124,7 +124,7 @@
 					<?php $allsum = $sum + $sum1 + $sum2 ?>
 					<tr>
 						<th colspan="7" class="text-right">GRAND TOTAL</th>
-						<th class="text-right"><?php echo e(number_format((int)$allsum,2)); ?></th>
+						<th class="text-right"><?php echo e(number_format((float)$allsum,2)); ?></th>
 					</tr>
 
 				</table>
